requirements.txt
pandas>=2.1.0,<4
numpy>=1.24.0,<3
openai==1.30.0
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.0
rich==13.7.0
